+++ packages/blocks/src/page-block/edgeless/tool-controllers/note-tool.ts
@@ -1,6 +1,4 @@
 import type { PointerEventState } from '@blocksuite/block-std';
-import { DisposableGroup } from '@blocksuite/global/utils';
-import { Overlay } from '@blocksuite/phasor';
 
 import {
   type EdgelessTool,
@@ -8,140 +6,133 @@ import {
   queryCurrentMode,
 } from '../../../__internal__/index.js';
 import { noop } from '../../../__internal__/index.js';
-import type { EdgelessPageBlockComponent } from '../edgeless-page-block.js';
 import {
   DEFAULT_NOTE_WIDTH,
-  NOTE_OVERLAY_CORNER_RADIUS,
-  NOTE_OVERLAY_DARK_BACKGROUND_COLOR,
-  NOTE_OVERLAY_HEIGHT,
-  NOTE_OVERLAY_LIGHT_BACKGROUND_COLOR,
   NOTE_OVERLAY_OFFSET_X,
   NOTE_OVERLAY_OFFSET_Y,
-  NOTE_OVERLAY_STOKE_COLOR,
-  NOTE_OVERLAY_TEXT_COLOR,
-  NOTE_OVERLAY_WIDTH,
 } from '../utils/consts.js';
 import { addNote, type NoteOptions } from '../utils/note.js';
+import { NoteOverlay } from '../utils/tool-overlay.js';
 import { EdgelessToolController } from './index.js';
 
-class NoteOverlay extends Overlay {
-  x = 0;
-  y = 0;
-  text = '';
-  globalAlpha = 0;
-  themeMode = 'light';
-  _edgeless: EdgelessPageBlockComponent;
-  _disposables!: DisposableGroup;
-  _lastViewportX: number;
-  _lastViewportY: number;
+// class NoteOverlay extends Overlay {
+//   x = 0;
+//   y = 0;
+//   text = '';
+//   globalAlpha = 0;
+//   themeMode = 'light';
+//   _edgeless: EdgelessPageBlockComponent;
+//   _disposables!: DisposableGroup;
+//   _lastViewportX: number;
+//   _lastViewportY: number;
 
-  private _getOverlayText(text: string): string {
-    return text[0].toUpperCase() + text.slice(1);
-  }
+//   private _getOverlayText(text: string): string {
+//     return text[0].toUpperCase() + text.slice(1);
+//   }
 
-  constructor(edgeless: EdgelessPageBlockComponent) {
-    super();
-    this._edgeless = edgeless;
-    this._disposables = new DisposableGroup();
-    this._lastViewportX = edgeless.surface.viewport.viewportX;
-    this._lastViewportY = edgeless.surface.viewport.viewportY;
-    this._disposables.add(
-      this._edgeless.slots.viewportUpdated.on(() => {
-        // get current viewport position
-        const currentViewportX = this._edgeless.surface.viewport.viewportX;
-        const currentViewportY = this._edgeless.surface.viewport.viewportY;
-        // calculate position delta
-        const deltaX = currentViewportX - this._lastViewportX;
-        const deltaY = currentViewportY - this._lastViewportY;
-        // update overlay current position
-        this.x += deltaX;
-        this.y += deltaY;
-        // update last viewport position
-        this._lastViewportX = currentViewportX;
-        this._lastViewportY = currentViewportY;
-        // refresh to show new overlay
-        this._edgeless.surface.refresh();
-      })
-    );
-    this._disposables.add(
-      this._edgeless.slots.edgelessToolUpdated.on(edgelessTool => {
-        // when change note child type, update overlay text
-        if (edgelessTool.type === 'note') {
-          this.text = this._getOverlayText(edgelessTool.tip);
-        }
-      })
-    );
-  }
+//   constructor(edgeless: EdgelessPageBlockComponent) {
+//     super();
+//     this._edgeless = edgeless;
+//     this._disposables = new DisposableGroup();
+//     this._lastViewportX = edgeless.surface.viewport.viewportX;
+//     this._lastViewportY = edgeless.surface.viewport.viewportY;
+//     this._disposables.add(
+//       this._edgeless.slots.viewportUpdated.on(() => {
+//         // get current viewport position
+//         const currentViewportX = this._edgeless.surface.viewport.viewportX;
+//         const currentViewportY = this._edgeless.surface.viewport.viewportY;
+//         // calculate position delta
+//         const deltaX = currentViewportX - this._lastViewportX;
+//         const deltaY = currentViewportY - this._lastViewportY;
+//         // update overlay current position
+//         this.x += deltaX;
+//         this.y += deltaY;
+//         // update last viewport position
+//         this._lastViewportX = currentViewportX;
+//         this._lastViewportY = currentViewportY;
+//         // refresh to show new overlay
+//         this._edgeless.surface.refresh();
+//       })
+//     );
+//     this._disposables.add(
+//       this._edgeless.slots.edgelessToolUpdated.on(edgelessTool => {
+//         // when change note child type, update overlay text
+//         if (edgelessTool.type === 'note') {
+//           this.text = this._getOverlayText(edgelessTool.tip);
+//         }
+//       })
+//     );
+//   }
 
-  override render(ctx: CanvasRenderingContext2D): void {
-    ctx.globalAlpha = this.globalAlpha;
-    // Draw the overlay rectangle
-    ctx.strokeStyle = NOTE_OVERLAY_STOKE_COLOR;
-    ctx.fillStyle =
-      this.themeMode === 'light'
-        ? NOTE_OVERLAY_LIGHT_BACKGROUND_COLOR
-        : NOTE_OVERLAY_DARK_BACKGROUND_COLOR;
-    ctx.lineWidth = 4;
-    ctx.beginPath();
-    ctx.moveTo(this.x + NOTE_OVERLAY_CORNER_RADIUS, this.y);
-    ctx.lineTo(
-      this.x + NOTE_OVERLAY_WIDTH - NOTE_OVERLAY_CORNER_RADIUS,
-      this.y
-    );
-    ctx.quadraticCurveTo(
-      this.x + NOTE_OVERLAY_WIDTH,
-      this.y,
-      this.x + NOTE_OVERLAY_WIDTH,
-      this.y + NOTE_OVERLAY_CORNER_RADIUS
-    );
-    ctx.lineTo(
-      this.x + NOTE_OVERLAY_WIDTH,
-      this.y + NOTE_OVERLAY_HEIGHT - NOTE_OVERLAY_CORNER_RADIUS
-    );
-    ctx.quadraticCurveTo(
-      this.x + NOTE_OVERLAY_WIDTH,
-      this.y + NOTE_OVERLAY_HEIGHT,
-      this.x + NOTE_OVERLAY_WIDTH - NOTE_OVERLAY_CORNER_RADIUS,
-      this.y + NOTE_OVERLAY_HEIGHT
-    );
-    ctx.lineTo(
-      this.x + NOTE_OVERLAY_CORNER_RADIUS,
-      this.y + NOTE_OVERLAY_HEIGHT
-    );
-    ctx.quadraticCurveTo(
-      this.x,
-      this.y + NOTE_OVERLAY_HEIGHT,
-      this.x,
-      this.y + NOTE_OVERLAY_HEIGHT - NOTE_OVERLAY_CORNER_RADIUS
-    );
-    ctx.lineTo(this.x, this.y + NOTE_OVERLAY_CORNER_RADIUS);
-    ctx.quadraticCurveTo(
-      this.x,
-      this.y,
-      this.x + NOTE_OVERLAY_CORNER_RADIUS,
-      this.y
-    );
-    ctx.closePath();
-    ctx.stroke();
-    ctx.fill();
+//   override render(ctx: CanvasRenderingContext2D): void {
+//     ctx.globalAlpha = this.globalAlpha;
+//     // Draw the overlay rectangle
+//     ctx.strokeStyle = NOTE_OVERLAY_STOKE_COLOR;
+//     ctx.fillStyle =
+//       this.themeMode === 'light'
+//         ? NOTE_OVERLAY_LIGHT_BACKGROUND_COLOR
+//         : NOTE_OVERLAY_DARK_BACKGROUND_COLOR;
+//     ctx.lineWidth = 4;
+//     ctx.beginPath();
+//     ctx.moveTo(this.x + NOTE_OVERLAY_CORNER_RADIUS, this.y);
+//     ctx.lineTo(
+//       this.x + NOTE_OVERLAY_WIDTH - NOTE_OVERLAY_CORNER_RADIUS,
+//       this.y
+//     );
+//     ctx.quadraticCurveTo(
+//       this.x + NOTE_OVERLAY_WIDTH,
+//       this.y,
+//       this.x + NOTE_OVERLAY_WIDTH,
+//       this.y + NOTE_OVERLAY_CORNER_RADIUS
+//     );
+//     ctx.lineTo(
+//       this.x + NOTE_OVERLAY_WIDTH,
+//       this.y + NOTE_OVERLAY_HEIGHT - NOTE_OVERLAY_CORNER_RADIUS
+//     );
+//     ctx.quadraticCurveTo(
+//       this.x + NOTE_OVERLAY_WIDTH,
+//       this.y + NOTE_OVERLAY_HEIGHT,
+//       this.x + NOTE_OVERLAY_WIDTH - NOTE_OVERLAY_CORNER_RADIUS,
+//       this.y + NOTE_OVERLAY_HEIGHT
+//     );
+//     ctx.lineTo(
+//       this.x + NOTE_OVERLAY_CORNER_RADIUS,
+//       this.y + NOTE_OVERLAY_HEIGHT
+//     );
+//     ctx.quadraticCurveTo(
+//       this.x,
+//       this.y + NOTE_OVERLAY_HEIGHT,
+//       this.x,
+//       this.y + NOTE_OVERLAY_HEIGHT - NOTE_OVERLAY_CORNER_RADIUS
+//     );
+//     ctx.lineTo(this.x, this.y + NOTE_OVERLAY_CORNER_RADIUS);
+//     ctx.quadraticCurveTo(
+//       this.x,
+//       this.y,
+//       this.x + NOTE_OVERLAY_CORNER_RADIUS,
+//       this.y
+//     );
+//     ctx.closePath();
+//     ctx.stroke();
+//     ctx.fill();
 
-    // Draw the overlay text
-    ctx.fillStyle = NOTE_OVERLAY_TEXT_COLOR;
-    let fontSize = 16;
-    ctx.font = `${fontSize}px Arial`;
-    ctx.textAlign = 'left';
-    ctx.textBaseline = 'middle';
+//     // Draw the overlay text
+//     ctx.fillStyle = NOTE_OVERLAY_TEXT_COLOR;
+//     let fontSize = 16;
+//     ctx.font = `${fontSize}px Arial`;
+//     ctx.textAlign = 'left';
+//     ctx.textBaseline = 'middle';
 
-    // measure the width of the text
-    // if the text is wider than the rectangle, reduce the maximum width of the text
-    while (ctx.measureText(this.text).width > NOTE_OVERLAY_WIDTH - 20) {
-      fontSize -= 1;
-      ctx.font = `${fontSize}px Arial`;
-    }
+//     // measure the width of the text
+//     // if the text is wider than the rectangle, reduce the maximum width of the text
+//     while (ctx.measureText(this.text).width > NOTE_OVERLAY_WIDTH - 20) {
+//       fontSize -= 1;
+//       ctx.font = `${fontSize}px Arial`;
+//     }
 
-    ctx.fillText(this.text, this.x + 10, this.y + NOTE_OVERLAY_HEIGHT / 2);
-  }
-}
+//     ctx.fillText(this.text, this.x + 10, this.y + NOTE_OVERLAY_HEIGHT / 2);
+//   }
+// }
 export class NoteToolController extends EdgelessToolController<NoteTool> {
   readonly tool = <NoteTool>{
     type: 'note',
@@ -256,8 +247,6 @@ export class NoteToolController extends EdgelessToolController<NoteTool> {
 
     // if mouse is in viewport and move, update overlay pointion and show overlay
     if (this._noteOverlay.globalAlpha === 0) this._noteOverlay.globalAlpha = 1;
-    // this._noteOverlay.text = this._getOverlayText();
-    // this._noteOverlay.themeMode = queryCurrentMode();
     const [x, y] = this._surface.viewport.toModelCoord(e.x, e.y);
     this._updateOverlayPosition(x, y);
   }
